Reorganizing buff flags, adding some helpful information
